chore(design): move assets out of components
