+++ design/components/content-block.stories.tsx
@@ -1,8 +1,8 @@
 import { Story } from '@storybook/react/types-6-0';
 import React from 'react';
 
-import begeleiding from './assets/begeleiding.png';
-import duurzaamheid from './assets/duurzaamheid.png';
+import begeleiding from '../assets/begeleiding.png';
+import duurzaamheid from '../assets/duurzaamheid.png';
 import { ContentBlock, ContentBlockProps } from './content-block';
 import { RICH_TEXT } from './data';
